fix SSE42 and x86_64
 lib/x86/mem_searchrn.c
 lib/x86/my_bitops.c
 lib/x86/strchrnul.c
 lib/x86/strlen.c
 lib/x86/strncasecmp_a.c
 lib/x86/strnlen.c
 lib/x86/strnpcpy.c
	finaly got my dirty hands on a Core i7
	       M E   W A N T !!!!1!!!11!!eleven *snif*
	so fix our SSE4.2 stuff, some things worked different than
	 thought, they take drugs at Intel...
	new ideas for the other stuff

 lib/x86/x86_features.h
	complete the feature list
	add AMDs mirrored flags, so you don't get naked numbers on
	 "print all flags"

 G2UDP.c
	wrong print spec for 64 Bit
 lib/bitfield_rle.c
	wrong type for 64 Bit register size
 lib/x86/aes.c
	no, it wont compile without Opt. on 64 Bit, so fix it prop.

 lib/strchr.c
	make a strchr wrapper around strchrnul
 lib/module.make
	add it to make

 lib/my_bitops.h
	add a strlitcmp, maybe wrong atm with a memcpy (memcpy can read
	 the full length, even if it would be dump if differences are
	 found before...)

 Makefile
	add other processors
	add -msse4 foo till gcc nows the right processor name

+++ G2UDP.c
@@ -148,7 +148,7 @@ static void *G2UDP_loop(void *param)
 	union combo_addr from, to;
 	int answer_fd = -1;
 
-	my_snprintf(d_hold.data, d_hold.limit, OUR_PROC " UDP %i", (int)param);
+	my_snprintf(d_hold.data, d_hold.limit, OUR_PROC " UDP %ti", (intptr_t)param);
 	g2_set_thread_name(d_hold.data);
 
 	while(worker.keep_going)
@@ -247,7 +247,7 @@ void *G2UDP(void *param)
 {
 	static pthread_t *helper;
 // TODO: get number of helper threads
-	unsigned int num_helper = 1, i;
+	size_t num_helper = 1, i;
 
 	poll_me[0].fd = *((int *)param);
 	logg(LOGF_INFO, "UDP:\t\tOur SockFD -> %d\tMain SockFD -> %d\n", poll_me[0].fd, *(((int *)param)-1));
@@ -299,7 +299,7 @@ void *G2UDP(void *param)
 	for(i = 0; i < num_helper; i++)
 	{
 		if(pthread_create(&helper[i], &server.settings.t_def_attr, G2UDP_loop, (void *)(i + 1))) {
-			logg_errnod(LOGF_WARN, "starting UDP helper threads, will run with %u", i);
+			logg_errnod(LOGF_WARN, "starting UDP helper threads, will run with %zu", i);
 			num_helper = i;
 			break;
 		}
